refactor(apps/web): update import paths for playback and audio utilities
- Refactored import paths for playback sync and audio metadata utilities to follow the new directory structure.
- Ensured consistency across components and tests by updating references to the new `playback/sync` and `audio` directories.
- Improved code organization and maintainability by aligning import statements with the updated file structure.
