+++ apps/web/src/components/library/tracks/bulk/useBulkTrackUpload.test.ts
@@ -1,16 +1,16 @@
-import { extractCoverFromAudioFile } from '@/lib/audio-metadata';
-import { cleanFilenameToTitle } from '@/lib/clean-audio-filename.ts';
+import { extractCoverFromAudioFile } from '@/lib/audio/audio-metadata';
+import { cleanFilenameToTitle } from '@/lib/audio/clean-audio-filename';
 import { albumBuilder, artistBuilder } from '@repo/testing/builders';
 import { customRenderHook } from '@repo/testing/web';
 import { act, waitFor } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { useBulkTrackUpload } from './useBulkTrackUpload';
 
-vi.mock('@/lib/audio-metadata', () => ({
+vi.mock('@/lib/audio/audio-metadata', () => ({
   extractCoverFromAudioFile: vi.fn(),
 }));
 
-vi.mock('@/lib/clean-audio-filename.ts', () => ({
+vi.mock('@/lib/audio/clean-audio-filename', () => ({
   cleanFilenameToTitle: vi.fn((name) => name.replace('.mp3', '')),
 }));
